Will extract duplicate test helpers into a shared scope
