+++ ModularFeedTDDCaseStudy/ModularFeedTDDCaseStudyTests/FeedCache/Helpers/FeedCacheTestHelpers.swift
@@ -1,0 +1,32 @@
+//
+//  FeedCacheTestHelpers.swift
+//  ModularFeedTDDCaseStudyTests
+//
+//  Created by Clint M on 4/2/19.
+//  Copyright © 2019 Clint M. All rights reserved.
+//
+
+import Foundation
+import ModularFeedTDDCaseStudy
+
+func uniqueImage() -> FeedImage {
+    return FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
+}
+
+func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
+    let models = [uniqueImage(), uniqueImage()]
+    let local = models.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.url) }
+    return (models, local)
+}
+
+extension Date {
+    
+    // if crash, developer error is the cause
+    func adding(days: Int) -> Date {
+        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
+    }
+    
+    func adding(seconds: TimeInterval) -> Date {
+        return self + seconds
+    }
+}
